feat(encoding): introduce serde_generic_array
