+++ encoding/src/lib.rs
@@ -10,7 +10,7 @@ pub mod serde_byte_array;
 pub use serde::{de, ser};
 pub use serde_bytes;
 pub use serde_cbor::{error, from_reader, from_slice, tags, to_vec, to_writer};
-pub mod generic_array;
+pub mod serde_generic_array;
 
 pub use self::bytes::*;
 pub use self::cbor::*;
